refactor(InputGetter.kt, InputParser.kt): 두 파일을 input 패키지 내에 InputManager.kt로 통합
