+++ src/main/kotlin/lotto/input/InputValidator.kt
@@ -1,0 +1,4 @@
+package lotto.input
+
+object InputValidator {
+}
