+++ src/test/kotlin/lotto/InputManagerTest.kt
@@ -1,0 +1,13 @@
+package lotto
+
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
+
+class InputManagerTest {
+    @Test
+    fun `정상적인 구매 금액 입력 테스트`() {
+
+    }
+
+    // TODO: 추가 기능 구현에 따른 테스트 코드 작성
+}
